Create separate files for extern spec rewriters
