+++ prusti-specs/src/extern_spec_rewriter/mod.rs
@@ -1,0 +1,6 @@
+//! Parsing of `#[extern_spec]` attributed structures
+
+pub mod impls;
+pub mod mods;
+pub mod traits;
+mod common;
